refactor: modularize routing
