+++ packages/manager/modules/metrics/src/dashboard/dashboard.routing.js
@@ -1,0 +1,16 @@
+import MetricsDashboardCtrl from './metrics-dashboard.controller';
+
+import dashboardTemplate from './metrics-dashboard.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('metrics.detail.dashboard', {
+    url: '/dashboard',
+    views: {
+      metricsContent: {
+        template: dashboardTemplate,
+        controller: MetricsDashboardCtrl,
+        controllerAs: '$ctrl',
+      },
+    },
+  });
+};
