+++ packages/manager/modules/metrics/src/details/index.js
@@ -14,20 +14,16 @@ import { Environment } from '@ovh-ux/manager-config';
 
 import ovhManagerServerSidebar from '@ovh-ux/manager-server-sidebar';
 import ngOvhCloudUniverseComponents from '@ovh-ux/ng-ovh-cloud-universe-components';
-import ovhManagerDashboardChartPie from '../dashboard/chart-pie';
 
-import MetricsDashboardCtrl from '../dashboard/metrics-dashboard.controller';
-import MetricsDetailCtrl from './metrics-detail.controller';
-import MetricsHeaderCtrl from '../header/metrics-header.controller';
+import dashboard from '../dashboard/dashboard.module';
+import token from '../token/token.module';
+import platform from '../platform/platform.module';
+
 import MetricService from './metrics.service';
 import FormatSiFilter from './format-si.filter';
 import routing from './routing';
 
-import '../dashboard/metrics-dashboard.less';
 import '../platform/metrics-platform.less';
-import '../token/metrics-token.less';
-import '../token/add/metrics-token-add.less';
-import '../token/preview/metrics-token-preview.less';
 
 const moduleName = 'ovhManagerMetricsDetails';
 
@@ -42,7 +38,9 @@ angular
     'pascalprecht.translate',
     'ui.bootstrap',
     'ui.router',
-    ovhManagerDashboardChartPie,
+    dashboard,
+    token,
+    platform,
     ovhManagerServerSidebar,
   ])
   .config(routing)
@@ -52,9 +50,6 @@ angular
       $qProvider.errorOnUnhandledRejections(false);
     },
   )
-  .controller('MetricsDashboardCtrl', MetricsDashboardCtrl)
-  .controller('MetricsDetailCtrl', MetricsDetailCtrl)
-  .controller('MetricsHeaderCtrl', MetricsHeaderCtrl)
   .service('MetricService', MetricService)
   .filter('formatSi', FormatSiFilter)
   .run(/* @ngTranslationsInject:json ./translations */);
