--- packages/manager/modules/metrics/src/header/constants.js
@@ -1,7 +1,0 @@
-export const GUIDE = {
-  FR: 'https://docs.ovh.com/fr/',
-};
-
-export default {
-  GUIDE,
-};
