--- packages/manager/modules/metrics/src/header/metrics-header.html
@@ -1,61 +1,0 @@
-<header class="oui-header">
-    <div class="oui-header__container">
-        <div class="oui-header__content">
-            <oui-spinner
-                data-size="m"
-                data-ng-if="MetricsDetailCtrl.loadingService"
-            ></oui-spinner>
-            <h1
-                class="oui-header__title"
-                data-ng-if="!MetricsDetailCtrl.loadingService"
-                data-ng-bind="MetricsDetailCtrl.service.description || MetricsDetailCtrl.serviceName"
-            ></h1>
-            <h2
-                class="oui-header__description"
-                data-ng-if="!MetricsDetailCtrl.loadingService && MetricsDetailCtrl.service.description !== MetricsDetailCtrl.serviceName"
-                data-ng-bind="MetricsDetailCtrl.serviceName"
-            ></h2>
-        </div>
-        <oui-guide-menu
-            data-text="{{ ::MetricsHeaderCtrl.guides.title }}"
-            placement="end"
-        >
-            <oui-guide-menu-group
-                data-ng-repeat="section in MetricsHeaderCtrl.guides.sections track by section.title"
-                data-label="{{ ::section.title }}"
-            >
-                <oui-guide-menu-item
-                    data-ng-repeat="guideItem in section.list track by guideItem.url"
-                    data-href="{{ ::guideItem.url }}"
-                    data-external="guideItem.external"
-                    ><span data-ng-bind="::guideItem.name"></span>
-                </oui-guide-menu-item>
-            </oui-guide-menu-group>
-            <oui-guide-menu-divider></oui-guide-menu-divider>
-            <oui-guide-menu-item
-                data-href="{{ ::MetricsHeaderCtrl.guides.footer.url }}"
-                data-external="MetricsHeaderCtrl.guides.footer.external"
-                ><span
-                    data-ng-bind="::MetricsHeaderCtrl.guides.footer.name"
-                ></span>
-            </oui-guide-menu-item>
-        </oui-guide-menu>
-    </div>
-    <oui-header-tabs>
-        <oui-header-tabs-item
-            data-state="metrics.detail.dashboard"
-            data-state-params="{ serviceName: MetricsHeaderCtrl.serviceName }"
-            ><span data-translate="metrics_tab_dashboard"></span>
-        </oui-header-tabs-item>
-        <oui-header-tabs-item
-            data-state="metrics.detail.token"
-            data-state-params="{ serviceName: MetricsHeaderCtrl.serviceName }"
-            ><span data-translate="metrics_tab_token"></span>
-        </oui-header-tabs-item>
-        <oui-header-tabs-item
-            data-state="metrics.detail.platform"
-            data-state-params="{ serviceName: MetricsHeaderCtrl.serviceName }"
-            ><span data-translate="metrics_tab_platform"></span>
-        </oui-header-tabs-item>
-    </oui-header-tabs>
-</header>
