+++ packages/manager/modules/metrics/src/platform/platform.routing.js
@@ -1,0 +1,15 @@
+import controller from './metrics-platform.controller';
+import template from './metrics-platform.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('metrics.detail.platform', {
+    url: '/platform',
+    views: {
+      metricsContent: {
+        template,
+        controller,
+        controllerAs: 'MetricsPlatformCtrl',
+      },
+    },
+  });
+};
