+++ packages/manager/modules/metrics/src/token/metrics-token.html
@@ -1,4 +1,4 @@
-<div class="metrics__token" data-ui-view="metricsSubContent">
+<div class="metrics__token" data-ui-view="metricsContent">
     <div class="metrics__row">
         <h3 class="oui-heading_3" data-translate="metrics_token_title"></h3>
         <p
